ENH: implement 'start-bluesky-httpserver' endpoint for starting the server
